add item in cart in baza de date next : trebuie sa repari view-ul de my cart.

+++ src/main/java/dao/ProductsDAO.java
@@ -8,5 +8,5 @@ public interface ProductsDAO {
     public List<ProductsEntity> findProducts();
     public List<ProductsEntity> findProductsByCategory(String category);
     public List<String> findAllCategory();
-    public ProductsEntity findProductById(int productId);
+    public ProductsEntity findProductById(Long productId);
 }
