+++ src/main/java/dto/UserDTO.java
@@ -3,7 +3,7 @@ package dto;
 import javax.persistence.Column;
 
 public class UserDTO {
-    private int id;
+    private Long id;
 
     private String userName;
 
@@ -15,12 +15,11 @@ public class UserDTO {
 
     private String address;
 
-
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
